StepRecurrence Local Handler
Instead of passing back form data to GoalStepForm I've added a intermediate handler for the data. This is because the combo boxes pass back the key of the input not the value.

This means once the recurrence form calls the submission handler the key and val pair is destroyed, so best to pass the val instead.